common/requirements.txt
gcloud>=0.18.3
google-api-python-client>=2.30.0
google-cloud-bigquery>=2.30.1
google-cloud-bigquery[pandas]
google-cloud-secret-manager>=2.7.0
gspread>=5.1.1
numpy>=1.17.0
oauth2client>=4.1.3
pandas>=0.25.3
pandas-gbq>=0.14.1
protobuf>=3.19.1
simplejson>=3.17.6
pandas-gbq==0.14.1
pytest==7.1.2
pytest-mock==3.7.0
instaloader==4.9
spacy==3.3.0
spacy-langdetect==0.1.2
spacy-legacy==3.0.9
spacy-loggers==1.0.2
aiohttp==3.7.0
praw==7.5.0
transformers==4.19.2
dynaconf==3.1.9
twint==2.1.20
aiohttp==3.7.0

